isolate jwt claims fixed: import cycle not allowed in test
